CRC32 is supposed to be an unsigned int
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1548937 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/archivers/sevenz/Archive.java
@@ -27,7 +27,7 @@ class Archive {
     /// Whether each particular packed streams has a CRC.
     BitSet packCrcsDefined;
     /// CRCs for each packed stream, valid only if that packed stream has one.
-    int[] packCrcs;
+    long[] packCrcs;
     /// Properties of solid compression blocks.
     Folder[] folders;
     /// Temporary properties for non-empty files (subsumed into the files array later).
